[pkg/orchestrator/redact] Use orchestrator build tag
