+++ pkg/orchestrator/redact/data_scrubber_test.go
@@ -9,16 +9,9 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	v1 "k8s.io/api/core/v1"
 
 	scrubberpkg "github.com/DataDog/datadog-agent/pkg/util/scrubber"
 )
-
-func BenchmarkNoRegexMatching1(b *testing.B)        { benchmarkMatching(1, b) }
-func BenchmarkNoRegexMatching10(b *testing.B)       { benchmarkMatching(10, b) }
-func BenchmarkNoRegexMatching100(b *testing.B)      { benchmarkMatching(100, b) }
-func BenchmarkNoRegexMatching1000(b *testing.B)     { benchmarkMatching(1000, b) }
-func BenchmarkRegexMatchingCustom1000(b *testing.B) { benchmarkMatchingCustomRegex(1000, b) }
 
 // https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go
 // store the result to a package level variable
@@ -26,61 +19,6 @@ func BenchmarkRegexMatchingCustom1000(b *testing.B) { benchmarkMatchingCustomReg
 //
 //goland:noinspection ALL
 var avoidOptimization bool
-
-//goland:noinspection ALL
-var avoidOptContainer v1.Container
-
-func benchmarkMatching(nbContainers int, b *testing.B) {
-	containersBenchmarks := make([]v1.Container, nbContainers)
-	containersToBenchmark := make([]v1.Container, nbContainers)
-	c := v1.Container{}
-
-	scrubber := NewDefaultDataScrubber()
-	for _, testCase := range getScrubCases() {
-		containersToBenchmark = append(containersToBenchmark, testCase.input)
-	}
-	for i := 0; i < nbContainers; i++ {
-		containersBenchmarks = append(containersBenchmarks, containersToBenchmark...)
-	}
-	b.ResetTimer()
-
-	b.Run("simplified", func(b *testing.B) {
-		for n := 0; n < b.N; n++ {
-			for _, c := range containersBenchmarks {
-				scrubContainer(&c, scrubber)
-			}
-		}
-	})
-	avoidOptContainer = c
-}
-
-func benchmarkMatchingCustomRegex(nbContainers int, b *testing.B) {
-	var containersBenchmarks []v1.Container
-	var containersToBenchmark []v1.Container
-	c := v1.Container{}
-
-	customRegs := []string{"pwd*", "*test"}
-	scrubber := NewDefaultDataScrubber()
-	scrubber.AddCustomSensitiveRegex(customRegs)
-
-	for _, testCase := range getScrubCases() {
-		containersToBenchmark = append(containersToBenchmark, testCase.input)
-	}
-	for i := 0; i < nbContainers; i++ {
-		containersBenchmarks = append(containersBenchmarks, containersToBenchmark...)
-	}
-
-	b.ResetTimer()
-	b.Run("simplified", func(b *testing.B) {
-		for n := 0; n < b.N; n++ {
-			for _, c := range containersBenchmarks {
-				scrubContainer(&c, scrubber)
-			}
-		}
-	})
-
-	avoidOptContainer = c
-}
 
 func TestMatchSimpleCommand(t *testing.T) {
 	cases := setupSensitiveCmdLines()
